app/vlselect/logsql: return VL-Request-Duration-Seconds HTTP header in response to HTTP querying APIs
This header contains the information about the duration for the query at the server
until the first response byte is sent to the client.
It doesn't include the time needed for sending the request to the server
and the time needed for reading the response from the server.

This duration will be displayed at the built-in web UI and the Grafana plugin for VictoriaLogs.